+++ src/Brave.cpp
@@ -1,21 +1,18 @@
 #include "RPJ.hpp"
-#include "Easter.hpp"
+#include "Brave.hpp"
 
-Easter::Easter() {
+Brave::Brave() {
 	config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
 
 	configParam(PARAM_FC, 20.f, 20480.f, 1000.f, "fc"," Hz");
 	configParam(PARAM_CVFC, 0.f, 1.0f, 0.0f, "CV FC");
 	configParam(PARAM_Q, 0.707f, 20.0f, 0.707f, "Q");
-	configParam(PARAM_CVQ, 0.f, 1.0f, 0.0f, "CV Q");
-	configParam(PARAM_DRY, 0.f, 1.0f, 0.0f, "DRY");
-	configParam(PARAM_WET, 0.f, 1.0f, 1.0f, "WET");
-	configParam(PARAM_UP, 0.0, 1.0, 0.0);
-	configParam(PARAM_DOWN, 0.0, 1.0, 0.0);
-	afp.algorithm = filterAlgorithm::kResonA;
+	configParam(PARAM_BOOSTCUT_DB, -20.f, 20.f, 0.f, "dB","Boost/Cut");
+	configParam(PARAM_CVB, 0.f, 1.0f, 0.0f, "CV Q");
+	afp.algorithm = filterAlgorithm::kNCQParaEQ;
 }
 
-void Easter::process(const ProcessArgs &args) {
+void Brave::process(const ProcessArgs &args) {
 
 	if (outputs[OUTPUT_MAIN].isConnected()) {
 		audioFilter.setSampleRate(args.sampleRate);
@@ -23,25 +20,22 @@ void Easter::process(const ProcessArgs &args) {
 		float cvfc = 1.f;
 		if (inputs[INPUT_CVFC].isConnected())
 			cvfc = inputs[INPUT_CVFC].getVoltage();
-	
+
 		float cvq = 1.f;
 		if (inputs[INPUT_CVQ].isConnected())
 			cvq = inputs[INPUT_CVQ].getVoltage();
+	
+		float cvb = 1.f;
+		if (inputs[INPUT_CVB].isConnected())
+			cvb = inputs[INPUT_CVB].getVoltage();
  	
 		afp.fc = params[PARAM_FC].getValue() * rescale(cvfc,-10,10,0,1);
 		afp.Q = params[PARAM_Q].getValue() * rescale(cvq,-10,10,0,1);
-		afp.dry = params[PARAM_DRY].getValue();
-		afp.wet = params[PARAM_WET].getValue();
-	
-		if (upTrigger.process(rescale(params[PARAM_UP].getValue(), 1.f, 0.1f, 0.f, 1.f))) 
-			afp.algorithm = filterAlgorithm::kResonB;	
-		if (downTrigger.process(rescale(params[PARAM_DOWN].getValue(), 1.f, 0.1f, 0.f, 1.f))) {
-		afp.algorithm = filterAlgorithm::kResonA;
-		}
+		afp.boostCut_dB = params[PARAM_BOOSTCUT_DB].getValue() * rescale(cvb,-10,10,0,1);
 
 		afp.strAlgorithm = audioFilter.filterAlgorithmTxt[static_cast<int>(afp.algorithm)];
 		audioFilter.setParameters(afp);
-		
+
 		float out = audioFilter.processAudioSample(inputs[INPUT_MAIN].getVoltage());
 		outputs[OUTPUT_MAIN].setVoltage(out);
 	}
@@ -63,11 +57,11 @@ struct buttonMinSmall : SvgSwitch  {
 	}
 };
 
-struct EasterModuleWidget : ModuleWidget {
-	EasterModuleWidget(Easter* module) {
+struct BraveModuleWidget : ModuleWidget {
+	BraveModuleWidget(Brave* module) {
 
 		setModule(module);
-		setPanel(APP->window->loadSvg(asset::plugin(pluginInstance, "res/Easter.svg")));
+		setPanel(APP->window->loadSvg(asset::plugin(pluginInstance, "res/Brave.svg")));
 
 		addChild(createWidget<ScrewSilver>(Vec(0, 0)));
 		addChild(createWidget<ScrewSilver>(Vec(box.size.x - 15, 365)));
@@ -76,13 +70,8 @@ struct EasterModuleWidget : ModuleWidget {
 
 		{
 			ATitle * title = new ATitle(box.size.x);
-			title->setText("EASTER");
+			title->setText("BRAVE");
 			addChild(title);
-		}
-		{
-			FilterNameDisplay * fnd = new FilterNameDisplay(Vec(39,30));
-			fnd->module = module;
-			addChild(fnd);
 		}
 		{
 			RPJTextLabel * tl = new RPJTextLabel(Vec(1, 50));
@@ -95,13 +84,8 @@ struct EasterModuleWidget : ModuleWidget {
 			addChild(tl);
 		}
 		{
-			RPJTextLabel * tl = new RPJTextLabel(Vec(58, 190));
-			tl->setText("DRY");
-			addChild(tl);
-		}
-		{
-			RPJTextLabel * tl = new RPJTextLabel(Vec(5, 190));
-			tl->setText("WET");
+			RPJTextLabel * tl = new RPJTextLabel(Vec(1, 170));
+			tl->setText("BOOST/CUT");
 			addChild(tl);
 		}
 		{
@@ -115,19 +99,17 @@ struct EasterModuleWidget : ModuleWidget {
 			addChild(tl);
 		}
 
-		addInput(createInput<PJ301MPort>(Vec(10, 300), module, Easter::INPUT_MAIN));
-		addOutput(createOutput<PJ301MPort>(Vec(55, 320), module, Easter::OUTPUT_MAIN));
+		addInput(createInput<PJ301MPort>(Vec(10, 300), module, Brave::INPUT_MAIN));
+		addOutput(createOutput<PJ301MPort>(Vec(55, 320), module, Brave::OUTPUT_MAIN));
 		
-		addParam(createParam<buttonMinSmall>(Vec(5,45),module, Easter::PARAM_DOWN));
-		addParam(createParam<buttonPlusSmall>(Vec(76,45),module, Easter::PARAM_UP));
-		addParam(createParam<RoundBlackKnob>(Vec(8, 80), module, Easter::PARAM_FC));
-		addInput(createInput<PJ301MPort>(Vec(55, 82), module, Easter::INPUT_CVFC));
-		addParam(createParam<RoundBlackKnob>(Vec(8, 140), module, Easter::PARAM_Q));
-		addInput(createInput<PJ301MPort>(Vec(55, 142), module, Easter::INPUT_CVQ));	
-		addParam(createParam<RoundBlackKnob>(Vec(8, 225), module, Easter::PARAM_WET));
-		addParam(createParam<RoundBlackKnob>(Vec(55, 225), module, Easter::PARAM_DRY));
+		addParam(createParam<RoundBlackKnob>(Vec(8, 80), module, Brave::PARAM_FC));
+		addInput(createInput<PJ301MPort>(Vec(55, 82), module, Brave::INPUT_CVFC));
+		addParam(createParam<RoundBlackKnob>(Vec(8, 140), module, Brave::PARAM_Q));
+		addInput(createInput<PJ301MPort>(Vec(55, 142), module, Brave::INPUT_CVQ));
+		addParam(createParam<RoundBlackKnob>(Vec(8, 200), module, Brave::PARAM_BOOSTCUT_DB));
+		addInput(createInput<PJ301MPort>(Vec(55, 202), module, Brave::INPUT_CVB));	
 	}
 
 };
 
-Model * modelEaster = createModel<Easter, EasterModuleWidget>("Easter");
+Model * modelBrave = createModel<Brave, BraveModuleWidget>("Brave");